refactor: moving compiled code to lib dir
modified outDir on tsconfig
changed package.json main field
